# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StellarSettle API",
    "version": "1.0.0",
    "description": "API specification for StellarSettle backend including Soroban contract interactions."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "RevokeKYCRequest": {
        "type": "object",
        "required": [
          "userId",
          "reviewerId"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "description": "User whose approval is being withdrawn."
          },
          "reviewerId": {
            "type": "string",
            "description": "Admin performing the revocation."
          },
          "revocationReason": {
            "type": "string",
            "description": "Recorded in the KYC audit log."
          }
        }
      },
...
      "AclEntry": {
        "type": "object",
        "properties": {
          "contractAddress": {
            "type": "string"
          },
          "permittedFunctions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "removed"
            ]
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          },
          "removedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastLedger": {
            "type": "string",
            "nullable": true
          },
          "lastTxHash": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AclLogEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contractAddress": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "enum": [
              "add",
              "remove"
            ]
          },
          "permittedFunctions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ledgerSequence": {
            "type": "string",
            "nullable": true
          },
          "txHash": {
            "type": "string",
            "nullable": true
          },
          "actor": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/admin/revoke-kyc": {
      "post": {
        "summary": "Revoke a user's KYC approval",
        "description": "Withdraws a previously granted KYC approval, returning the user to the pending state so they can be reviewed again. Only an approved user can be revoked. Requires the admin API key and an allow-listed source IP.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeKYCRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "KYC revoked"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "User is not currently approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/acl": {
      "get": {
        "summary": "Get the contract integration ACL",
        "description": "Returns the whitelisted contract integrations and the functions each may call, projected from ACLUpdated contract events. Cached for five minutes and invalidated whenever a new ACLUpdated event is indexed. Admin only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "contracts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AclEntry"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "cachedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "Admin access required": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "Authentication required": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/acl/log": {
      "get": {
        "summary": "Get the ACL add/remove history",
        "description": "Returns the append-only audit trail of ACL changes derived from ACLUpdated events, newest first. Admin only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AclLogEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "Admin access required": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "Authentication required": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "Bad Request": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}